requirements.txt
Django==3.2.4
psycopg2-binary>=2.8
python-environ
django-cors-headers
django-rest-framework
django-widget-tweaks
djangorestframework~=3.13.1
djangorestframework-jwt~=1.11.0
django-braces~=1.15.0
django-storages[dropbox]
dropbox~=11.30.0
cloudinary~=1.29.0
django-cloudinary-storage
django-cloudinary-storage[video]
django-redis
pillow
django-cors-headers
celery

redis~=4.3.1
environ~=1.0